Add an Audio Recorder Polyfill for iOS
Is a polyfill really all it takes? Nope! See follow-on commits for
updates to our audio code to get it running on iOS.

I pushed this early to start testing on staging ASAP. More details below
for the curious:

This is really hard to test locally. In insecure contexts, including
serving over a non-SSL'd local devserver, `navigator.mediaDevices` is
`undefined`. Going to <localIP>:9000/speak from the devserver itself
doesn't even work. So, testing microphone support locally on a phone is
a non-starter. See
https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getUserMedia#Security
for further details.